requirements.txt
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
alembic==1.14.0
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.4.0
asteroid-filterbanks==0.4.0
async-timeout==4.0.3
attrs==24.2.0
audioread==3.0.1
av==12.3.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
coloredlogs==15.0.1
colorlog==6.9.0
contourpy==1.3.0
ctranslate2==4.5.0
cycler==0.12.1
dataclasses-json==0.6.7
decorator==5.1.1
distro==1.9.0
docopt==0.6.2
docx2txt==0.8
einops==0.8.0
exceptiongroup==1.2.2
faiss==1.8.0
fastapi==0.114.1
faster-whisper==1.0.3
filelock==3.16.0
flatbuffers==24.3.25
fonttools==4.54.1
frozenlist==1.4.1
fsspec==2024.9.0
greenlet==3.1.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.2
huggingface-hub==0.24.7
humanfriendly==10.0
HyperPyYAML==1.2.2
idna==3.8
importlib_resources==6.4.5
Jinja2==3.1.4
jiter==0.5.0
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
julius==0.2.7
kiwisolver==1.4.7
langchain==0.2.16
langchain-community==0.2.16
langchain-core==0.2.39
langchain-huggingface==0.0.3
langchain-openai==0.1.23
langchain-text-splitters==0.2.4
langsmith==0.1.118
lazy_loader==0.4
librosa==0.10.2.post1
lightning==2.4.0
lightning-utilities==0.11.8
llvmlite==0.43.0
Mako==1.3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.22.0
matplotlib==3.9.2
mdurl==0.1.2
mkl-service==2.4.0
mkl_fft @ file:///C:/b/abs_f55mv94vyg/croot/mkl_fft_1725370278455/work
mkl_random @ file:///C:/b/abs_21ydbzdu8d/croot/mkl_random_1725370276095/work
mpmath==1.3.0
msgpack==1.1.0
multidict==6.1.0
mypy-extensions==1.0.0
networkx==3.2.1
nltk==3.9.1
numba==0.60.0
numpy @ file:///C:/b/abs_c1ywpu18ar/croot/numpy_and_numpy_base_1708638681471/work/dist/numpy-1.26.4-cp39-cp39-win_amd64.whl#sha256=f284633067fde4bb078a09085a1c9e99a3218c129f5312b7b3d76c9340442a25
omegaconf==2.3.0
onnxruntime==1.19.2
openai==1.45.0
optuna==4.0.0
orjson==3.10.7
packaging @ file:///C:/b/abs_c3vlh0z4jw/croot/packaging_1720101866539/work
pandas==2.2.3
pdf2image==1.17.0
pillow==10.4.0
platformdirs==4.3.6
pooch==1.8.2
primePy==1.3
protobuf==5.28.3
pyannote.audio==3.1.1
pyannote.core==5.0.0
pyannote.database==5.1.0
pyannote.metrics==3.2.1
pyannote.pipeline==3.0.1
pycparser==2.22
pydantic==2.9.1
pydantic_core==2.23.3
Pygments==2.18.0
pyparsing==3.2.0
pypdf==5.0.1
pyreadline3==3.5.4
pytesseract==0.3.13
python-dateutil==2.9.0.post0
pytorch-lightning==2.4.0
pytorch-metric-learning==2.7.0
pytz==2024.2
PyYAML==6.0.2
regex==2024.9.11
requests==2.32.3
rich==13.9.4
ruamel.yaml==0.18.6
ruamel.yaml.clib==0.2.12
safetensors==0.4.5
scikit-learn==1.5.2
scipy==1.13.1
semver==3.0.2
sentence-transformers==3.1.0
sentencepiece==0.2.0
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
soundfile==0.12.1
soxr==0.5.0.post1
speechbrain==1.0.2
SQLAlchemy==2.0.34
starlette==0.38.5
sympy==1.13.1
tabulate==0.9.0
tenacity==8.5.0
tensorboardX==2.6.2.2
threadpoolctl==3.5.0
tiktoken==0.7.0
tokenizers==0.19.1
torch==2.5.1+cu124
torch-audiomentations==0.11.1
torch_pitch_shift==1.2.5
torchaudio==2.5.1+cu124
torchmetrics==1.5.2
torchvision==0.20.1+cu124
tqdm==4.66.5
transformers==4.44.2
typer==0.13.0
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
uvicorn==0.30.6
whisperx==3.1.6
yarl==1.11.1
zipp==3.21.0
